requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chromadb==0.3.23
click==8.1.7
clickhouse-connect==0.8.17
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
docx2txt==0.9
duckdb==1.2.1
fastapi==0.115.11
filelock==3.16.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
fpdf==1.7.2
frozenlist==1.6.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
hf-xet==1.1.2
hnswlib==0.8.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.62
langchain-text-splitters==0.3.8
langsmith==0.3.43
lz4==4.4.4
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.41.0
networkx==3.4.2
numpy==2.2.3
openai==1.82.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.6
plotly==5.24.1
posthog==4.2.0
preswald==0.1.42
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydeck==0.9.1
PyMuPDF==1.26.0
PyMySQL==1.1.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==76.0.0
simpy==4.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.1
streamlit==1.45.1
sympy==1.14.0
tenacity==9.0.0
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.0
torch-geometric==2.6.1
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.28.0
watchfiles==1.0.5
websockets==10.4
Werkzeug==3.1.3
yarl==1.20.0
zstandard==0.23.0
